feat(output): Support outputting the glossary to a file
Output needs to be split if the glossary needs to outputted to a file.
Informative output which is not the actual end result (i.e. the glossary)
still needs to be written to the expected output channel (i.e. the cli).

In addition, the output channel is actually split between cli output (in
the form of a PrintStream) and the maven log (in case the maven plugin is
used).

Three destinations (ToFile, ToPrintStream and ToLog) encapsulate these
different output channels.

The PrintableGlossary deals with outputting the correct data to the correct
channels because it knows what needs to be outputted and thus also knows the
context in which it needs to be outputted (info/target).