chore(qulice): (WIP) Add qulice to police code quality
Still a Work In Progress because the FindBugs part of
Qulice is not yet able to process a JDK 11 project. For
now, Qulice is still disabled in normal builds.

To use qulice: mvn install -Pqulice

This required quite a lot of changes, but resolving the
PMD and checkstyle violations improved the code and
consistency.