feat(types): Support annotations and enums (#14)
The `@GlossaryTerm` annotation could only be processed when it 
annotated a class or interface. From now on, the `@Term` annotation 
replaces the old `@GlossaryTerm` and adds the ability to transform 
annotation and enum declarations into glossary definitions.

The `Classifier` class no longer captured each of these types, so it has 
been aptly renamed to `Type`. 

Processing order used to be:
 files -> classifiers (class/interface) -> definitions

Processing order is now:
 files -> types (classifier/enum/annotation) -> definitions

Resolves #12