chore(releases): Replace maven-release-plugin
Simply use the `${revision}`, which is available since maven version
3.5.0, and tag the commit with the revision.

Versioning (for releases) can now be done using:
`mvn clean install -Prelease -Drevision=<version>`

For deployments to ossrh/central this means that the pom has to be
cleaned. This is done using the flatten-plugin.

See https://maven.apache.org/maven-ci-friendly.html or
https://axelfontaine.com/blog/dead-burried.html for more information.