fix(type): Fail compilation on warnings (#16)
* feat(errors): Fail compilation on warnings

From now on compilation will fail when the compiler raises warnings. 
This should improve the quality of the code.

* fix(type): Fix warning unchecked type

The compiler gave a warning that the constructor call from within the
other constructor in TypeFromJavaparser was unchecked. Apparantly the
type expected was different from the type given. The extra type
parameter information has been added to deal with the warning.