chore(from-java): Rename java project to from-java
To allow other tools to be added to the git repo, the java project that
extracts a glossary/cantisfile from a java codebase is being moved to
its own subfolder. Additionally, the project is renamed to fit its
purpose: from-java.

Cantis can now be a monorepo containing a multitude of tools, like
to-html, to-confluence, to-react and to-angular.